♻️ add drag and drop image in offer
